Switch to pid messages from game state
Basically, I wanted to implement some automated smoke testing with bots, and instead of writing them as websocket clients, which is fine but seems like an overkill for this project, I decided to make them simple GenServer pids that join games.

So? That means game state and game master shouldn’t assume they are interacting with websocket channels, hence the need to decouple some of the broadcast topics.

Yes, I could’ve kept it and simply register as a listener to the PubSub around that topic, but why? Why involve this complicated PubSub system from Phoenix? Better isolate ourselves and keep the footprint minimal.